Modifications to existing code to run pointer chase benchmarks, scripts for pointer chase benchmarks

+++ hw/block/femu/computation.c
@@ -1,8 +1,8 @@
-int count_bits(char *buf)
+uint64_t count_bits(char *buf)
 {
         int i, j;
         char c;
-        int count =0;
+        uint64_t count =0;
 
         for (i = 0 ; i < 4096 ; i++) {
                 c = buf[i];
@@ -15,8 +15,8 @@ int count_bits(char *buf)
         return count;
 }
 
-int get_disk_pointer(char *buf)
+uint64_t get_disk_pointer(char *buf)
 {
-	int *p = buf;
+	uint64_t *p = buf;
 	return p[0];
 }
